+++ src/app.ts
@@ -8,7 +8,7 @@ import compression from 'compression';
 import Routes from './interfaces/routes.interface';
 import errorMiddleware from './middlewares/error.middleware';
 import { logger, stream } from './utils/logger';
-import { connectMongoMemoryDb } from './utils/mongoMemoryDB';
+import { connectMongoDb } from './utils/mongoDB';
 
 class App {
   public app: express.Application;
@@ -27,7 +27,7 @@ class App {
   }
 
   public connectDatabase() {
-    connectMongoMemoryDb();
+    connectMongoDb();
   }
 
   public listen() {
